Fix some clang compiler warnings

+++ include/Field_C.h
@@ -50,13 +50,13 @@ extern void DestroyGSimpleField(void* field, int coords);
 extern void DestroyKSimpleField(void* field, int coords);
 extern void DestroyNSimpleField(void* field, int coords);
 
-extern void KMeansInitTree(void* field, double* centers, long npatch, int d, int coords);
-extern void KMeansInitRand(void* field, double* centers, long npatch, int d, int coords);
-extern void KMeansInitKMPP(void* field, double* centers, long npatch, int d, int coords);
-extern void KMeansRun(void* field, double* centers, long npatch, int max_iter, double tol,
+extern void KMeansInitTree(void* field, double* centers, int npatch, int d, int coords);
+extern void KMeansInitRand(void* field, double* centers, int npatch, int d, int coords);
+extern void KMeansInitKMPP(void* field, double* centers, int npatch, int d, int coords);
+extern void KMeansRun(void* field, double* centers, int npatch, int max_iter, double tol,
                       int alt, int d, int coords);
-extern void KMeansAssign(void* field, double* centers, long npatch,
+extern void KMeansAssign(void* field, double* centers, int npatch,
                          long* patches, long n, int d, int coords);
 
-extern void SelectPatch(int patch, double* centers, long npatch, double* x, double* y, double* z,
+extern void SelectPatch(int patch, double* centers, int npatch, double* x, double* y, double* z,
                         long* use, long n);
